refactor: 💡 Move module sidebar to specific folder
