+++ apps/e-learning-app/src/app/components/moduleSidebar/Sidebar.js
